- Token 작업 중.

+++ DsDotNet/src/Engine/Engine.CodeGenCPU/Core.Extensions/Vertex, Call, Real.fs
@@ -175,7 +175,7 @@ module ConvertCpuVertex =
     type Real with
         member r.V = r.TagManager :?> RealVertexTagManager
 
-        member r.RealSEQ:uint64 = r.V.GetVertexTag(VertexTag.realSEQ).BoxedValue :?> uint64 
+        member r.RealSEQ:uint32 = r.V.GetVertexTag(VertexTag.realSEQ).BoxedValue :?> uint32
 
         member r.CoinSTContacts = r.Graph.Vertices.Select(getVMCoin).Select(fun f->f.ST)
         member r.CoinRTContacts = r.Graph.Vertices.Select(getVMCoin).Select(fun f->f.RT)
@@ -200,5 +200,5 @@ module ConvertCpuVertex =
 [<Extension>]
 type RealExt =
     [<Extension>]
-    static member GetRealSEQ(r:Real):uint64 = r.RealSEQ
+    static member GetRealSEQ(r:Real):uint32 = r.RealSEQ
